requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
einops>=0.7.0
scipy>=1.11.0

# VLM backbone
transformers>=4.45.0
accelerate>=0.25.0

# Fine-tuning
peft>=0.10.0
deepspeed>=0.14.0

# Data processing
pillow>=10.0.0
pyyaml>=6.0
huggingface-hub>=0.20.0

# Video decoding (DROID dataset)
# ffmpeg must be installed via conda: conda install -c conda-forge ffmpeg
